refactor IAM subscription method
respond to PR comments

respond to PR comments
